add arch test
