+++ src/modules/exchange-rate/application/interfaces/exchange-rate-notification.service.interface.ts
@@ -1,4 +1,4 @@
-import { ExchangeRate } from '../../../domain/entities/exchange-rate.entity';
+import { ExchangeRate } from '../../domain/entities/exchange-rate.entity';
 
 export interface ExchangeRateNotificationService {
   sendExchangeRateNotification(
