+++ src/modules/exchange-rate/infrastructure/notification/exchange-rate-email.service.ts
@@ -1,12 +1,12 @@
 import { Inject, Injectable } from '@nestjs/common';
 
+import { MailerService } from 'src/modules/mailer/application/interfaces/mailer.service.interface';
 import { ExchangeRateEmailComposerService } from 'src/modules/mailer/infrastructure/email/interfaces/exchange-rate-email-composer.service.interface';
-import { MailerService } from 'src/modules/mailer/infrastructure/email/interfaces/mailer.service.interface';
-import { TYPES as MAILER_TYPES } from 'src/modules/mailer/infrastructure/ioc';
+import { TYPES as MAILER_TYPES } from 'src/modules/mailer/ioc';
 
-import { ExchangeRateNotificationService } from './interfaces/exchange-rate-notification.service.interface';
+import { ExchangeRateNotificationService } from '../../application/interfaces/exchange-rate-notification.service.interface';
 import { ExchangeRate } from '../../domain/entities/exchange-rate.entity';
-import { TYPES as EXCHANGE_RATE_TYPES } from '../../infrastructure/ioc';
+import { TYPES as EXCHANGE_RATE_TYPES } from '../../ioc';
 
 @Injectable()
 export class ExchangeRateNotificationServiceImpl
