+++ src/modules/mailer/application/interfaces/mailer.service.interface.ts
@@ -1,4 +1,4 @@
-import { Email } from '../../../domain/entities/email.entity';
+import { Email } from '../../domain/entities/email.entity';
 
 export interface MailerService {
   sendEmail(email: Email): Promise<void>;
