+++ src/modules/subscription/subscription.module.ts
@@ -5,8 +5,8 @@ import { PrismaModule } from 'src/shared/infrastructure/prisma/prisma.module';
 import { CreateSubscriptionApplicationImpl } from './application/create-subscription.application';
 import { SubscriptionController } from './controller/subscription.controller';
 import { SubscriptionServiceImpl } from './domain/services/subscription.service';
-import { TYPES } from './infrastructure/ioc/types';
 import { PrismaSubscriptionRepositoryImpl } from './infrastructure/repositories/prisma-subscription.repository';
+import { TYPES } from './ioc/types';
 
 const createSubscriptionApp = {
   provide: TYPES.applications.CreateSubscriptionApplication,
